Updated generator script

+++ engine/dlib/src/dmsdk/dlib/trig_lookup.h
@@ -68,7 +68,7 @@ namespace dmTrigLookup
      */
     inline float Sin(float radians)
     {
-        return Cos(radians - (float) M_PI_2);
+        return Cos(radians - M_PI_2);
     }
 
 }
